Home

CamiDiff.jl is a Julia package for one-dimensional finite-difference analysis.

Install

The package is installed using the Julia package manager

julia> using Pkg; Pkg.add("CamiDiff")

julia> using CamiDiff

Introduction

CamiDiff has been developed to analyze mathematical functions, provided by the user in tabulated form. The package is based on finite-difference analysis methods in one dimension.

Throughout the documentation $f(x)$ will be the 'function of interest' under investigation. In tabulated form we write $f[n]$. This form can be regarded as the result of the discretization of $f(x)$ on a Grid of $N$ points, addressable by the grid index $n = 1, ⋯ N$. The Grid can be linear (uniform) or non-linear as specified by a gridfunction - see Discretization.

The current implementation of CamiDiff was developed for real functions of a single variable. A set of four predefined gridfunctions is included in the package: 'exponential', 'truncated-exponential', 'linear' and 'polynomial'. These are restricted to the domain $[0, ∞)$. When applicable, we shall underline this restriction to the non-negative domain by using the variable $r$ rather than $x$, writing $f(r)$ rather than $f(x)$, with the implicit condition $r ≥ 0$.

Discretization

Mathematically, the discretization is based on the map $n ↦ x,$ which defines the discrete function

\[x[n] = s_0 * g(t[n]).\]

Here $g(t)$ is called the gridfunction and $s_0$ the scaling factor. The gridfunction is defined as a (generally nonlinear) function running through the origin: $g(0) = 0$. Its argument is the ticks function

\[t[n] ≡ (n−u) * h,\]

which is a discrete linear function, where $u$ is called the index base and $h$ the step size. Writing

\[f[n] = f(x[n]),\]

we recognize in $f[n]$ a discrete function representing the function $f(x)$ at position $x[n]$. This represents the tabulated function to be provided by the user.

NB. The discrete function $f[n]$ is defined on the grid of natural numbers, a uniform grid with unit step size. This uniformity greatly simplifies the numerical analysis. The stepsize of the ticks function, $h$, determines the coarseness of the grid. The results of a finite-difference calculation on a coarse grid will be less accurate than those on a fine grid, but the algorithm is identical, because the relevant finite-difference expansions only depend implicitly on $h$. Since Julia uses unit-based indexing ($u = 1$), the index convention implies $f[1] = f(0)$.

Grid

The Grid object is the backbone for numerical procedures on the real domain $[0, ∞)$. Its principal fields are grid.r, grid.r′ and grid.r′′. These are discrete functions of N elements representing the grid function and its first two derivatives. The function $f[n]$ is tabulated on this Grid and the function $r[n]$= grid.r represents the transformation by the gridfunction.

Once the Grid is specified, three Grid operations are at our disposal - see Documentation

grid_interpolation(f, rval, grid)

grid_differentiation(f, grid)

grid_integration(f, grid)

CamiDiff.Grid — Type

Grid(ID, name, T, N, r, r′, r′′, h, r0, epn, epw, k, p, polynom)

Type with fields:

.ID: grid identifer name (::Int)

.name: grid identifer name (::String)

.T: gridtypename (::Type)

.N: number of grid points (::Int)

.r: tabulated grid function (::Vector{T})

.r′: tabulated first derivative of grid function (::Vector{T})

.r′′: tabulated second derivative of grid function (::Vector{T})

.h : grid step multiplyer (::T)

.r0: grid scale factor (::T)

.epn: number of endpoints used for trapezoidal endpoint correction (must be odd) (::Int)

.epw: trapezoidal endpoint weights for n=1:epn (::Vector{Vector{T}})

.k: finite-difference order (::Int)

.p: only for truncated-exponential grid; truncation power (::Int)

.polynom: only for polynomial grid: polynom (::Vector{T})

The object Grid is best created with the function castGrid.

CamiDiff.castGrid — Method

castGrid(ID::Int, N::Int, T::Type; h=1, rmax=10, p=5, polynom=[0,1], epn=5, k=5, msg=false)
castGrid(name::String, N::Int, T::Type; h=1, rmax=10, p=5, polynom=[0,1], epn=5, k=5, msg=false)

Method to create a Grid object that covers the radial range [0, rmax] with N points.

ID = 1: exponential, ID = 2: truncated-exponential, ID = 3: linear (uniform) ID = 4: polynomial

Examples:

julia> grid = castGrid(1, 1000, Float64; h = 0.005, rmax = 10, msg=true);
Grid: exponential, Float64, rmax = 10.0, Ntot = 1000, h = 0.005, r0 = 0.0681789

julia> grid = castGrid("exponential", 1000, Float64; h = 0.005, rmax = 10, msg=true);
Grid: exponential, Float64, rmax = 10.0, Ntot = 1000, h = 0.005, r0 = 0.0681789

julia> grid = castGrid(2, 1000, Float64; h = 0.005, rmax = 10, p=5, msg=true);
Grid: truncated-exponential, Float64, rmax = 10.0, Ntot = 1000, p = 5, h = 0.005, r0 = 0.111

julia> grid = castGrid(3, 1000, Float64; h = 0.1, rmax = 10, msg=true);
Grid: linear (uniform), Float64, rmax = 10.0, Ntot = 1000, p = 1, h = 0.1, r0 = 0.1001

julia> grid = castGrid(4, 1000, Float64; h = 0.1, rmax = 10, polynom=[0,0,1], msg=true);
Grid: polynomial of degree 2, Float64, rmax = 10.0, Ntot = 1000, polynom = [0.0, 0.0, 1.0], h = 0.1, r0 = 0.001002

julia> r = grid.r[1:4]; println("r = ", r)
r = [0.0, 1.002003004005006e-5, 4.008012016020024e-5, 9.018027036045053e-5]

julia> r = grid.r[997:1000]; println("r = ", r)
r = [9.9400301202103, 9.96000004008012, 9.979990000010021, 10.0] # note the end of the range (r = rmax) 

julia> r′= grid.r′[1:4]; println("r′ = ", r′)
r′ = [0.0, 2.0040060080100123e-5, 4.008012016020025e-5, 6.012018024030035e-5]

julia> r′′= grid.r′′[1:4]; println("r′′ = ", r′′)
r′′ = [2.004006008010012e-5, 2.004006008010012e-5, 2.004006008010012e-5, 2.004006008010012e-5]

CamiDiff.gridfunction — Method

gridfunction(ID::Int, n::Int, T::Type; h=1, p=5, polynom=[0,1], deriv=0)

CamiDiff offers three internal grid functions:

ID = 1: exponential grid function,

\[ g(t) = \rm{exp}(t) - 1\]

ID = 2: truncated-exponential grid function of degree p (linear grid for p = 1),

\[ g(t) = t + \frac{1}{2}t^2 + ⋯ + \frac{1}{p!}t^p\]

ID = 3: linear grid function,

\[ g(t) = t\]

ID = 4: polynomial grid function of degree p = length(c)-1 defined by its CamiMath.polynom vector $c = [c_0, c_1,c_2,⋯\ c_p]$,

\[ g(t) = c_0 + c_1 t + c_2 t^2 + ⋯ + c_p t^p,\]

with $c_0 ≡ 0$ because, by definition, all gridfunctions run through the origin, $g(0) = 0$.

The actual Grid is given by

\[ r[n] = r_0 * g(t[n]),\]

where $t[n] = (n-1) * h$ is the ticks function for the unit-based indexing of Julia.

NB. All gridfunctions satisfy the properties $t[1] = 0$ and $r[1] = 0$.

Examples:

julia> h = 0.1; r0=1.0; N=4; T=Float64;

julia> r = [r0*gridfunction(1, n, T; h) for n=1:N]; println("r = ", r)
r = [0.0, 0.10517091807564771, 0.22140275816016985, 0.3498588075760032]

julia> r′= r0 .* [r0*gridfunction(1, n, T; h, deriv=1) for n=1:N]; println("r′= ", r′)
r′ = [0.1, 0.11051709180756478, 0.122140275816017, 0.13498588075760032]

julia> r′′= [r0*gridfunction(1, n, T; h, deriv=2) for n=1:N]; println("r′′= ", r′′)
r′′ = [0.010000000000000002, 0.011051709180756479, 0.012214027581601701, 0.013498588075760034]

julia> r = [r0*gridfunction(4, n, T; h, polynom=[0,0,1]) for n=1:N]; println("r = ", r)
r = [0.0, 0.010000000000000002, 0.04000000000000001, 0.09000000000000002]

julia> r′= [r0*gridfunction(4, n, T; h, polynom=[0,0,1], deriv=1) for n=1:N]; println("r′= ", r′)
r′= [0.0, 0.020000000000000004, 0.04000000000000001, 0.06000000000000001]

julia> r′′= [r0*gridfunction(4, n, T; h, polynom=[0,0,1], deriv=2) for n=1:N]; println("r′′= ", r′′)
r′′= [0.020000000000000004, 0.020000000000000004, 0.020000000000000004, 0.020000000000000004]

CamiDiff.gridtypename — Method

gridtypename(ID::Int)

Name corresponding to the Grid ID.

Example:

julia> gridtypename(2)
"truncated-exponential"

CamiDiff.gridtypeID — Method

gridtypeID(name::String)

ID corresponding to the gridtypename.

Example:

julia> gridtypeID("truncated-exponential")
2

CamiDiff.gridPos — Method

gridPos(rval::T, grid::Grid{T}) where T<:Number

The approximate grid position defined as the largest integer n satisfying the condition grid.r[n] < rval on the Grid.

Example:

Consider the exponential grid of 4 points defined by

julia> grid = castGrid("exponential", 4, Float64; h = 0.1, rmax = 2.0);

julia> println(grid.r)
[0.0, 0.6012192107114549, 1.265669197778149, 2.0]

The approximate grid position of the point r = 1.0 is n = 2 (larger than 0.6012192107114549 but smaller than 1.265669197778149).

julia> r = 1.0;

julia> n = gridPos(r, grid)
2

CamiDiff.fracPos — Method

fracPos(n::Int, rval::T, grid::Grid{T}; ϵ = 1e-8, k = 7) where T<:Real

Fractional grid offset with respect to Grid position n.

Example:

Consider the exponential grid of 4 points defined by

julia> grid = castGrid("exponential", 4, Float64; h = 0.1, rmax = 2.0);

julia> println(grid.r)
[0.0, 0.6012192107114549, 1.265669197778149, 2.0]

Te point r = 1.0 is located at approximate n n = 2, with fractional position Δn = 0.6120806373655796.

julia> r = 1.0;

julia> n = gridPos(r, grid)
2

julia> Δn = fracPos(n, r, grid);

julia> t = (n+Δn-1)*grid.h;

julia> grid.r0 * (exp(t)-1) ≈ r