Finite-difference methods

Finite differences

Consider the analytical function $f$ tabulated in forward order (growing index) at $n$ positions on a uniform grid.

Forward difference notation

In forward difference notation, the finite difference of two adjacent values on the grid is defined as

\[Δ f[n] = f[n+1]-f[n],\]

where $Δ$ is the forward difference operator. By a formal inversion procedure we find

\[f[n-1]=(1+Δ)^{-1}f[n]=(1-Δ+Δ^2-Δ^3+⋯)f[n],\]

where $Δ^k$ is the $k^{th}$-order forward difference defined as a weighted sum over the function values $f[n:n+k]$ (involving $k+1$ points),

\[Δ^k f[n] = c_{k}^kf[n] + c_{k-1}^kf[n+1] + ⋯ + f[n+k] = \sum_{j=0}^{k} c_{k-j}^k f[n+j].\]

The $k+1$ coefficients

\[c_{k-j}^{k}=(-1)^{k-j}\binom{k}{j}\]

are the summation weights (short: weights) which define the summation.

Backward difference notation

In backward difference notation, the finite difference of two adjacent values on the grid is defined as

\[∇ f[n] = f[n]-f[n-1],\]

where $∇$ is the backward difference operator. By a formal inversion procedure we find

\[f[n+1]=(1-∇)^{-1}f[n]=(1+∇+∇^2+∇^3+⋯)f[n],\]

where $∇^k$ is the $k^{th}$-order backward difference defined as a weighted sum over the function values tabulated in backward order, $f[n:-1:n-k]$ (involving $k+1$ points),

\[∇^k f[n] = f[n] + c_1^kf[n-1] + ⋯ + c_k^kf[n-k] = \sum_{j=0}^{k} c_j^kf[n-j],\]

where the $k+1$ coefficients

\[c_{j}^{k}=(-1)^{j}\binom{k}{j}\]

are the summation weights (short: weights) which define the summation. Note the special cases $c_{0}^{k}≡1$, $c_{k}^{k}≡(-1)^{k}$ and the symmetry relation

\[c_{k-j}^k=(-1)^k c_j^k.\]

Coefficients:

fdiff_weight(k,j) $→ c_j^k=(-1)^j\binom{k}{j}$

CamiDiff.fdiff_weight — Method

fdiff_weight(k::Int, j::Int)

Finite difference weight coefficient,

\[c_{j}^{k}=(-1)^{k+j}\binom{k}{j}.\]

Example:

c(k,j) = fdiff_weight(k,j)

[[c(k,j) for j=0:k] for k=0:3] == [[1], [1, -1], [1, -2, 1], [1, -3, 3, -1]]
  true

[[c(k,k-j) for j=0:k] for k=0:3] == [[1], [-1, 1], [1, -2, 1], [-1, 3, -3, 1]]
  true

Finite difference expansions

Finite-difference calculus builds on the finite-difference expansion.

Forward difference notation

In terms of forward differences the expansion takes the form

\[\sum_{p=0}^{\infty}α_{p}Δ^{p}f[n] =\sum_{p=0}^{k}α_{p}Δ^{p}f[n]+⋯.\]

A finite-difference expansion truncated at order $k$ is defined by $k+1$ finite-difference expansion coefficients, represented by the vector $α = [α_{0},⋯\ α_{k}]$. It takes some bookkeeping to rewrite the expansion as a weighted sum over the $k+1$ function values in forward tabulated form $f[n:n+k]$. Substituting the finite difference expression for $Δ^k$, we obtain

\[\sum_{p=0}^{k}α_{p}Δ^{p}f[n] =\sum_{p=0}^{k}α_{p}\sum_{j=0}^{p}c_{p-j}^{p}f[n+j] =\sum_{j=0}^{k}\sum_{p=j}^{k}α_{p}c_{p-j}^{p}f[n+j] =\sum_{j=0}^{k}F_{j}^{k}f[n+j],\]

where the weighted summation is defined by the weights

\[F_{j}^{k}=\sum_{p=j}^{k}α_{p}c_{p-j}^{p} =\sum_{p=j}^{k}(-1)^{p+j}\binom{p}{j}α_{p},\]

with $j=0,⋯\ k$. In inner product form the expansion becomes

\[\sum_{p=0}^{k}α_{p}Δ^{p}f[n] =\sum_{j=0}^{k}F_{j}^{k}f[n+j] =F^{k} \cdot f[n:n+k],\]

where $F^k ≡ [F_0^k,⋯\ F_k^k]$.

\[f[n:n+k] = \left[\begin{array}{c} f[n]\\ \vdots\\ f[n+k] \end{array}\right].\]

Coefficients:

fdiff_expansion_weights(polynom, fwd, reg) $→ F^k ≡ [F_0^k,⋯\ F_k^k]$,

where polynom is the CamiMath.polynom vector $α ≡ [α_0,⋯\ α_k]$. This polynom has to be supplied by the user to define the expansion under consideration. Some common cases are:

interpolation expansion: fdiff_interpolation_expansion_polynom(ξ, k, fwd)

differentiation expansion: fdiff_differentiation_expansion_polynom(ξ, k, fwd)

Adams-Bashford expansion: fdiff_adams_bashford_expansion_polynom(k, fwd)

Adams-Moulton expansion: fdiff_adams_moulton_expansion_polynom(k, fwd)

Backward difference notation

In terms of backward differences the expansion takes the form

\[\sum_{p=0}^{\infty}β_{p}∇^{p}f[n]=\sum_{p=0}^{k}β_{p}∇^{p}f[n]+⋯.\]

In this case the $k^{th}$- order finite-difference expansion is defined by the vector $β = [β_{0},⋯\ β_{k}]$. The expansion can written as weighted sum over the $k+1$ function values in backward tabulated form $f[n:-1:n-k]$. Substituting the finite difference expression for $∇^k$, we obtain

\[\sum_{p=0}^{k}β_{p}∇^{p}f[n] =\sum_{p=0}^{k}β_{p}\sum_{j=0}^{p}c_{j}^{p}f[n-j] =\sum_{j=0}^{k}\sum_{p=j}^{k}β_{p}c_{j}^{p}f[n-j] =\sum_{j=0}^{k}B_{j}^{k}f[n-j],\]

where the weights are given by

\[B_{j}^{k}=\sum_{p=j}^{k}β_{p}c_{j}^{p} =\sum_{p=j}^{k}(-1)^{j}\binom{p}{j}β_{p},\]

with $j=0,⋯\ k$. In inner product form the expansion becomes

\[\sum_{p=0}^{k}β_{p}∇^{p}f[n] =\sum_{j=0}^k B_j^k f[n-j] =\bar{B}^k \cdot f[n-k:n],\]

where the weights vector $\bar{B}^{k} ≡ [B_k^k,⋯\ B_0^k]$ contains the weights in backward order.

In general there is no simple symmetry relation between $B^k$ and $F^k$.

Coefficients:

fdiff_expansion_weights(polynom, bwd, rev) $→ \bar{B}^{k} ≡ [B_k^k,⋯\ B_0^k]$,

where polynom is the CamiMath.polynom vector $β ≡ [β_0,⋯\ β_k]$. This polynom has to be supplied by the user to define the expansion under consideration. Some common cases are:

interpolation expansion: fdiff_interpolation_expansion_polynom(ξ, k, bwd)

differentiation expansion: fdiff_differentiation_expansion_polynom(ξ, k, bwd)

Adams-Bashford expansion: fdiff_adams_bashford_expansion_polynom(k, bwd)

Adams-Moulton expansion: fdiff_adams_moulton_expansion_polynom(k, bwd)

CamiDiff.fdiff_expansion_weights — Function

fdiff_expansion_weights(polynom[, notation=bwd[, ordering=rev]])

Expansion weights corresponding to the expansion coefficients CamiMath.polynom of a finite difference expansion.

Forward difference notation (notation = fwd)

Weight vector $F^k ≡ [F_k^k,⋯\ F_0^k]$ corresponding to the expansion coefficients $α ≡ [α_0^k,⋯\ α_k^k]$ of the $k^{th}$-order forward-difference expansion,

\[\sum_{p=0}^{k}α_{p}Δ^{p}f[n] =\sum_{j=0}^{k}F_{j}^{k}f[n+j] =F^{k} \cdot f[n:n+k],\]

where $f[n:n+k]$ are elements of the analytic function $f$ tabulated in forward order.

fdiff_expansion_weights(α, fwd, reg) $→ F^k ≡ [F_0^k,⋯\ F_k^k]$,

where $α ≡ [α_0,⋯\ α_k]$ has to be supplied in combination with fwd to indicate that the weights must be evaluated in forward-difference notation.

Backward difference notation (notation = bwd)

Weight vector $\bar{B}^{k} ≡ [B_k^k,⋯\ B_0^k]$ corresponding to the expansion coefficients $β ≡ [β_0,⋯\ β_k]$ of the $k^{th}$-order backward-difference expansion,

\[\sum_{p=0}^{k}β_{p}∇^{p}f[n] =\sum_{j=0}^{k}B_{j}^kf[n-j] =\bar{B}^k \cdot f[n-k:n].\]

where $f[n-k:n]$ are elements of the analytic function $f$ tabulated in forward order.

fdiff_expansion_weights(β, bwd, rev) $→ \bar{B}^{k} ≡ [B_k^k,⋯\ B_0^k]$,

where $β ≡ [β_0,⋯\ β_k]$ has to be supplied in combination with bwd to indicate that the weights must be evaluated in backward-difference notation.

Example:

Consider the expansions,

\[f[n-1]=(1+Δ)^{-1}f[n]=(1-Δ+Δ^2-Δ^3+⋯)f[n].\]

\[f[n+1]=(1-∇)^{-1}f[n]=(1+∇+∇^2+∇^3+⋯)f[n],\]

α = [1,-1,1,-1,1]
β = [1,1,1,1,1]
Fk = fdiff_expansion_weights(α, fwd, reg); println("Fk = $(Fk)")
  Fk = [5, -10, 10, -5, 1]

Bk = fdiff_expansion_weights(β, bwd, reg); println("Bk = $(Bk)")
  Bk = [5, -10, 10, -5, 1]

revFk = fdiff_expansion_weights(α, fwd, rev); println("revFk = $(revFk)")
  revFk = [1, -5, 10, -10, 5]

revBk = fdiff_expansion_weights(β, bwd, rev); println("revBk = $(revBk)")
  revBk = [1, -5, 10, -10, 5]

CamiDiff.fdiff_expansion — Function

fdiff_expansion(polynom, f[, notation=bwd])

Finite difference expansion of the analytical function f(x) tabulated in forward order (growing index) at $k+1$ positions on a uniform grid. The expansion coefficients are specified by the vector CamiMath.polynom. By default CamiMath.polynom are assumed to be in backward-difference notation (bwd). For CamiMath.polynom in forward-difference notation the third argument must be fwd.

Forward difference notation (notation = fwd)

\[\sum_{p=0}^{k}α_{p}Δ^{p}f[n] = F^{k} \cdot f[n:n+k],\]

where $f[n:n+k]$ are elements of the analytical function $f$ (tabulated in forward order) and $α ≡ [α_0,⋯\ α_k]$ is the vector CamiMath.polynom, which has to be supplied to define the forward-difference expansion. The corresponding weights vector $F^{k}$ is internally generated.

Backward difference notation (notation = bwd)

\[\sum_{p=0}^{k}β_{p}∇^{p}f[n] = \bar{B}^k \cdot f[n-k:n].\]

where $f[n-k:n]$ are elements of the analytical function $f$ (tabulated in forward order) and $β ≡ [β_0,⋯\ β_k]$ is the vector CamiMath.polynom, which has to be supplied to define the backward-difference expansion. The corresponding weights vector $\bar{B}^k$ is internally generated.

Examples:

Consider the function $f(x)=x^2$ and the expansions,

\[f(x-1)=(1+Δ)^{-1}=(1-Δ+Δ^2-Δ^3+⋯)f(x).\]

\[f(x+1)=(1-∇)^{-1}=(1+∇+∇^2+∇^3+⋯)f(x),\]

To fourth order (k=4) the forward- and backward-difference coefficient vectors are α=[1,-1,1,-1,1] and β=[1,1,1,1,1], respectively. We tabulate the function at $k+1$ points, f=[1,4,9,16,25].

α = [1,-1,1,-1,1]
Fk = fdiff_expansion_weights(α, fwd, reg); println("Fk = $(Fk)")
  Fk = [5, -10, 10, -5, 1]

β = [1,1,1,1,1]
revBk = fdiff_expansion_weights(β, bwd, rev); println("revBk = $(revBk)")
  revBk = [1, -5, 10, -10, 5]

f = [1,4,9,16,25]
o = fdiff_expansion(α, f, fwd); println("f[0] = $(o)")
  f[0] = 0

fdiff_expansion(α, f, fwd) == Fk ⋅ f == fdiff_interpolation(f, 0)
  true

o = fdiff_expansion(β, f, bwd); println("f[6] = $(o)")
  f[6] = 36

fdiff_expansion(β, f, bwd) == revBk ⋅ f == fdiff_interpolation(f, length(f)+1)
  true

In these cases the results are exact because the function is quadratic and the expansion is third order (based on the polynomial of $k^{th}$ degree running through the $k+1$ points of the tabulated function). Note the relation with fdiff_interpolation(f, v, k=3).