Home

CamiDiff.jl is a Julia package for one-dimensional finite-difference analysis.

Install

The package is installed using the Julia package manager

julia> using Pkg; Pkg.add("CamiDiff")

julia> using CamiDiff

Introduction

CamiDiff has been developped to study continuously differentiable functions, provided by the user in tabulated form. The package is based on the methods of finite-difference analysis in one dimension.

Throughout the documentation $f(x)$ will be our function of interest under investigation. The tabulated form can be regarded as the result of discretization of $f(x)$, onto a Grid of $N$ points, addressable by the gridindex $n = 1, ⋯ N$. The Grid can be linear or non-linear as specified by a gridfunction - see Discretization.

The current implementation of CamiDiff was developped for real functions of a single variable, using gridfunctions restricted to the domain $[0, ∞)$. A set of four predefined gridtypenames is included: 'exponential', 'quasi-exponential', 'linear' and 'polynomial'. To underline the restriction to the non-negative domain, we shall often use the variable $r$ rather than $x$, writing $f(r)$ rather than $f(x)$, with the implicit condition $r ≥ 0$.

Discretization

Mathematically, the discretization is based on the map $n ↦ x,$ which defines the discrete function

\[x[n] = s_0 * g(t[n]).\]

Here $g(t)$ is called the gridfunction and $s_0$ the scaling factor. The gridfunction is defined as a (generally nonlinear) function running through the origin: $g(0) = 0$. Its argument is the ticks function

\[t[n] ≡ (n−u) * h,\]

which is a discrete linear function, where $u$ is called the index base and $h$ the step size. Writing

\[f[n] = f(x[n]),\]

we recognize in $f[n]$ a discrete function representing the function $f(x)$ at position $x[n]$. This represents the tabulated that has to be provided by the user

Note that $h$ determines the coarseness of the Grid. The results of a finite-difference calculation on a coarse grid will be less accurate than those on a fine grid, but the algorithm is identical, because the relevant finite-difference expansions only depend on $h$ implicitely. Since Julia uses unit-based indexing ($u = 1$), the index convention implies $f[1] = f(0)$.

Grid

The Grid object is the backbone for the numerical procedure on a (generally) non-uniform grid. Its principal fields are grid.r, grid.r′ and grid.r′′ which are discrete functions of N elements representing the grid function and its first two derivatives.

Once the Grid is specified, three basic operations are at our disposal - see Application

grid_interpolation(f, rval, grid)

grid_differentiation(f, grid)

grid_integration(f, grid)

CamiDiff.Grid — Type

Grid(ID, name, T, N, r, r′, r′′, h, r0, epn, epw, k, p, polynom)

Type with fields:

.ID: grid identifer name (::Int)

.name: grid identifer name (::String)

.T: gridtypename (::Type)

.N: number of grid points (::Int)

.r: tabulated grid function (::Vector{T})

.r′: tabulated first derivative of grid function (::Vector{T})

.r′′: tabulated second derivative of grid function (::Vector{T})

.h : grid step multiplyer (::T)

.r0: grid scale factor (::T)

.epn: number of endpoints used for trapezoidal endpoint correction (must be odd) (::Int)

.epw: trapezoidal endpoint weights for n=1:epn (::Vector{Vector{T}})

.k: finite-difference order (::Int)

.p: only for quasi-exponential grid; truncation power (::Int)

.polynom: only for polynomial grid: polynom (::Vector{T})

The object Grid is best created with the function castGrid.

CamiDiff.castGrid — Method

castGrid(ID::Int, N::Int, T::Type; h=1, r0=1,  p=5, polynom=[0,1], epn=5, k=7, msg=false)
castGrid(name::String, N::Int, T::Type; h=1, r0=1, p=5, polynom=[0,1], epn=5, k=7, msg=false)

Method to create the Grid object

ID = 1: exponential, ID = 2: quasi-exponential, ID = 3: linear (uniform) ID = 4: polynomial

Examples:

julia> grid = castGrid(1, 1000, Float64; h = 0.005, r0 = 0.1, msg=true);
Grid: exponential, Float64, rmax = 14.7413, Ntot = 1000, h = 0.005, r0 = 0.1

julia> grid = castGrid("exponential", 1000, Float64; h = 0.005, r0 = 0.1, msg=true);
Grid: exponential, Float64, rmax = 14.7413, Ntot = 1000, h = 0.005, r0 = 0.1

julia> grid = castGrid(2, 1000, Float64; h = 0.005, r0 = 0.1, p=5, msg=true);
Grid: quasi-exponential, Float64, rmax = 9.04167, Ntot = 1000, p = 5, h = 0.005, r0 = 0.1

julia> grid = castGrid(3, 1000, Float64; h = 0.1, r0 = 0.1, msg=true);
Grid: linear (uniform), Float64, rmax = 10.0, Ntot = 1000, p = 1, h = 0.1, r0 = 0.1

julia> grid = castGrid(4, 1000, Float64; h = 0.1, r0 = 0.001, polynom=[0,0,1], msg=true);
Grid: polynomial of degree 2, Float64, rmax = 10.0, Ntot = 1000, polynom = [0.0, 0.0, 1.0], h = 0.1, r0 = 0.001

julia> grid.r[1:4]
4-element Vector{Float64}:
 [2.220446049250313e-16, 1.0000000000000003e-5, 4.000000000000001e-5, 9.000000000000003e-5]

julia> grid.r′[1:4]
4-element Vector{Float64}:
 [0.0, 2.0000000000000005e-5, 4.000000000000001e-5, 6.0000000000000015e-5]
 
julia> grid.r′′[1:4]
4-element Vector{Float64}:
 [2.0000000000000005e-5, 2.0000000000000005e-5, 2.0000000000000005e-5, 2.0000000000000005e-5] 

CamiDiff.gridfunction — Method

gridfunction(ID::Int, n::Int, h::T; p=5, polynom=[0,1], deriv=0) where T <: Real

CamiDiff offers three internal gridfunctions:

ID = 1: exponential grid function,

\[ g(t) = e^t - 1\]

ID = 2: quasi-exponential grid function of degree p (linear grid for p = 1),

\[ g(t) = t + \frac{1}{2}t^2 + ⋯ + \frac{1}{p!}t^p\]

ID = 3: linear grid function,

\[ g(t) = t\]

ID = 4: polynomial grid function of degree p = length(c)-1 defined by its CamiMath.polynom vector $c = [c_0, c_1,c_2,⋯\ c_p]$,

\[ g(t) = c_0 + c_1 t + c_2 t^2 + ⋯ + c_p t^p,\]

with $c_0 ≡ 0$ because, by definition, all grid functions run through the origin, $g(0) = 0$.

The actual grid is given by

\[ x[n] = r_0 * g(t[n]),\]

where $t[n] = (n-1) * h$ is the ticks function for the unit-based indexing of Julia.

NB. Note that $t[1] = 0$ and $x[1] = 0$ for all grid functions.

Examples:

julia> h = 0.1; r0=1.0; N=4;

julia> r = r0 .* [gridfunction(1, n-1, h) for n=1:N]
[0.0, 0.10517091807564771, 0.22140275816016985, 0.3498588075760032]

julia> r′ = r0 .* [gridfunction(1, n-1, h; deriv=1) for n=1:N]
[0.1, 0.11051709180756478, 0.122140275816017, 0.13498588075760032]

julia> r′′= r0 .* [gridfunction(1, n-1, h; deriv=2) for n=1:N]
[0.010000000000000002, 0.011051709180756479, 0.012214027581601701, 0.013498588075760034]

julia> r = r0 .* [gridfunction(4, n-1, h; polynom=[0,0,1]) for n=1:N]
[0.0, 0.010000000000000002, 0.04000000000000001, 0.09000000000000002]

julia> r′ = r0 .* [gridfunction(4, n-1, h; polynom=[0,0,1], deriv=1) for n=1:N]
[0.0, 0.020000000000000004, 0.04000000000000001, 0.06000000000000001]

julia> r′′= r0 .* [gridfunction(4, n-1, h; polynom=[0,0,1], deriv=2) for n=1:N]
[0.020000000000000004, 0.020000000000000004, 0.020000000000000004, 0.020000000000000004]

CamiDiff.gridtypename — Method

gridtypename(ID::Int)

Name corresponding to the Grid ID.

Example:

julia> gridtypename(2)
"quasi-exponential"

CamiDiff.gridtypeID — Method

gridtypeID(name::String)

ID corresponding to the gridtypename.

Example:

julia> gridtypeID("quasi-exponential")
2

CamiDiff.gridPos — Method

gridPos(rval::T, grid::Grid{T}) where T<:Number

The approximate grid position defined as the largest integer n satisfying the condition grid.r[n] < rval on the Grid.

Example:

julia> h = 0.1; r0 = 1.0;

julia> grid = castGrid(1, 4, Float64; h, r0);

julia> r = grid.r; println("r[3] = $(r[3])")
r[3] = 0.22140275816016985

julia> gridPos(0.222, grid)
3

CamiDiff.fracPos — Method

fracPos(n::Int, rval::T, grid::Grid{T}; ϵ = 1e-8, k = 7) where T<:Real

Fractional grid offset with respect to Grid position n.

Finite differences

Having discretized the analytic function $f(x)$, one defines finite differences. The forward difference of $f(x+h)$ and $f(x)$ is defined by

\[Δ f(x)=f(x+h)-f(x).\]

Here, $Δ$ is called the forward-difference operator. Likewise one defines backward differences with the backward-difference operator $∇$,

\[∇ f(x)=f(x)-f(x-h).\]

We first focus on forward differences. The derivative of $f(x)$ is given by

\[f^′(x)=\underset{h→0}{\mathrm{lim}}\,\frac{f(x+h)-f(x)}{h}=\underset{Δ x→0}{\mathrm{lim}}\,\frac{Δ f(x)}{Δ x},\]

where $h ≡ Δx ≥ 0$ is the difference interval. Introducing the differential operator, $f^′(x) ≡ Df(x)$, we have

\[D≡\frac{d}{dx}=\underset{Δ x→0}{\mathrm{lim}}\,\frac{Δ}{Δ x}=\underset{h→0}{\mathrm{lim}}\,\frac{Δ}{h}.\]

Translation operators

Forward difference notation

With regard to forward differences we rewrite the forward difference definition in the form of a forward translation,

\[f(x+h)=(1+Δ)f(x),\]

where $T≡(1+Δ)$ is the forward translation operator. This operator shifts the function over the infinitesimal interval $h$ to larger values of $x$. The translation operator can be expressed in terms of the differential operator as follows by Taylor expansion of $f(x)$ about the point $x$,

\[f(x± h)=(1± hD+\tfrac{1}{2}h^2D^2±\tfrac{1}{3!}h^3D^3+⋯)f(x)=e^{± hD}f(x).\]

Comparing the operator expression for the forward translation with the Taylor expansion we obtain, by formal inversion of the operator $T$, an operator identity for the inverted translation operator $T^{-1}$,

\[T≡(1+Δ)=e^{hD}\,\,\,⇒\,\,\,T^{-1}=e^{-hD}=(1+Δ)^{-1}.\]

With this procedure, the explicit dependence on $h$ can be replaced by an implicit dependence on $h$ through an expansion in powers of $Δ$ ,

\[f(x-h)=(1+Δ)^{-1}f(x)=(1-Δ+Δ^{2}-Δ^3+⋯)f(x).\]

By choosing the proper expansion order, $f(x-h)$ can be approximated to any desired level of accuracy.

Backward difference notation

Likewise, for backward differences, we rewrite the backward-difference definition in the form

\[f(x-h)=(1-∇)f(x),\]

where $B≡(1-∇)$ is the backward translation operator. Comparing this backward translation with the Taylor expansion we obtain, by formal inversion of the operator $B$, an operator identity for the forward translation operator $T$,

\[B≡(1-∇)=e^{-hD}=T^{-1}\,\,\,⇒\,\,\,T=e^{hD}=(1+∇)^{-1}.\]

Note how the backward translation operator was identified with the inverse forward translation operator, $B=T^{-1}$. When using backward differences, the explicit dependence on $h$ can be replaced by an implicit dependence on $h$ through an expansion in powers of $∇$,

\[f(x+h)=(1-∇)^{-1}f(x)=(1+∇+∇^{2}+∇^3+⋯)f(x).\]

By choosing the proper expansion order, $f(x+h)$ can be approximated to any desired level of accuracy.